(#6510) - respect ajax timeout when fetching _changes
Previously, fetching _changes used a timeout value defined in
opts.timeout. This is incorrect - that timeout value represents the
query parameter passed as part of the _changes call. Instead, use the
timeout specified in ajaxOpts.timeout.

PouchDB already has logic to create a buffer between the server side
timeout and the client timeout, providing a 5 second grace period
for the server to respond before the client will timeout. This change
makes that work as expected.